requirements.txt
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
django-simple-history==3.7.0
uuid6==2024.7.10
numpy==2.1.1
celery==5.4.0
redis==5.2.1
